Rename 'Tokenizer' as 'SimpleTokenizer'
